backend/requirements.txt
# Core web dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==20.1.0
python-multipart==0.0.6
httpx==0.25.1

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Configuration
pydantic==2.5.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
email-validator==2.1.0  # Required for Pydantic email field validation

# Vector databases
pymilvus==2.2.9  # Using a more stable version
marshmallow<3.20.0  # Specifying compatible marshmallow version
environs<9.0.0  # Specifying compatible environs version

# NLP and ML dependencies
scikit-learn==1.3.0  # For TF-IDF vectorization and ML utilities
numpy==1.24.3  # Needed for numerical operations
scipy==1.11.1  # Needed for sparse matrices and scientific computing

# spaCy for text processing
spacy==3.7.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# Testing
pytest==7.4.3

# Optional utilities
textblob==0.17.1
nltk==3.8.1
python-slugify